+++ src/main/kotlin/racingcar/model/CarCount.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.model
 
 data class CarCount(
     val count: Int
