+++ src/main/kotlin/racingcar/model/MovingStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.model
 
 fun interface MovingStrategy {
     fun movable(): Boolean
