+++ src/main/kotlin/racingcar/view/Output.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.view
+
+import racingcar.model.Cars
 
 class Output {
