+++ src/test/kotlin/racingcar/CarCountTest.kt
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.model.CarCount
 
 internal class CarCountTest {
