+++ src/test/kotlin/racingcar/RandomValueMovingStrategyTest.kt
@@ -2,6 +2,8 @@ package racingcar
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.model.RandomValueMovingStrategy
+
 internal class RandomValueMovingStrategyTest {
